+++ src/Form/CheckBox.cpp
@@ -217,5 +217,5 @@ CheckBoxControl::OnPaint(Canvas &canvas)
   canvas.Select(*cb_look.font);
   canvas.SetTextColor(state_look.text_color);
   canvas.SetBackgroundTransparent();
-  canvas.DrawText({canvas.GetHeight() + 2, 2}, caption.c_str());
+  canvas.DrawText({(int)canvas.GetHeight() + 2, 2}, caption.c_str());
 }
